Increase max aggregation to 256K
Without this change, NET ISOs had two sequences of 7 consecutive metadata chunks
and with this change, this became a single chunk of 102400 bytes.
While in theory this could reduce chunk reuse, I had not observed such reuse
in practice and the lower number of chunks will decrease overhead
of storage and DHT handling.